+++ wezterm-gui-subcommands/src/lib.rs
@@ -204,6 +204,19 @@ pub struct StartCommand {
     #[arg(long, verbatim_doc_comment)]
     pub position: Option<GuiPosition>,
 
+    /// Name of the multiplexer domain section from the configuration
+    /// to which you'd like to connect. If omitted, the default domain
+    /// will be used.
+    #[arg(long)]
+    pub domain: Option<String>,
+
+    /// When used with --domain, if the domain already has running panes,
+    /// wezterm will simply attach and will NOT spawn the specified PROG.
+    /// If you omit --attach when using --domain, wezterm will attach
+    /// AND then spawn PROG.
+    #[arg(long, requires = "domain")]
+    pub attach: bool,
+
     /// Instead of executing your shell, run PROG.
     /// For example: `wezterm start -- bash -l` will spawn bash
     /// as if it were a login shell.